Move FragmentKey into android package.
