+++ formula-android/src/main/java/com/instacart/formula/android/FeatureFactory.kt
@@ -1,7 +1,6 @@
 package com.instacart.formula.android
 
 import com.instacart.formula.fragment.FormulaFragment
-import com.instacart.formula.fragment.FragmentKey
 import com.instacart.formula.fragment.FragmentFlowStore
 
 /**
