+++ formula-android/src/main/java/com/instacart/formula/android/FragmentBindingBuilder.kt
@@ -3,7 +3,6 @@ package com.instacart.formula.android
 import com.instacart.formula.android.internal.FunctionUtils
 import com.instacart.formula.android.views.FragmentContractViewFactory
 import com.instacart.formula.fragment.FragmentContract
-import com.instacart.formula.fragment.FragmentKey
 import com.instacart.formula.android.internal.FeatureBinding
 import com.instacart.formula.integration.Integration
 import io.reactivex.rxjava3.core.Observable
