+++ formula-android/src/main/java/com/instacart/formula/fragment/BaseFormulaFragment.kt
@@ -1,5 +1,7 @@
 package com.instacart.formula.fragment
 
+import com.instacart.formula.android.FragmentKey
+
 interface BaseFormulaFragment<State> {
 
     fun getFragmentKey(): FragmentKey
