+++ formula-android/src/main/java/com/instacart/formula/fragment/FragmentContract.kt
@@ -1,6 +1,7 @@
 package com.instacart.formula.fragment
 
 import android.view.View
+import com.instacart.formula.android.FragmentKey
 
 /**
  * Responsible for providing a [FragmentComponent]
