+++ samples/todoapp/src/main/java/com/examples/todoapp/tasks/TaskListKey.kt
@@ -1,6 +1,6 @@
 package com.examples.todoapp.tasks
 
-import com.instacart.formula.fragment.FragmentKey
+import com.instacart.formula.android.FragmentKey
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
